backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
spacy==3.7.2
python-dotenv==1.0.0
groq==0.4.0
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6 